+++ src/fparser/common/tests/test_readfortran.py
@@ -103,7 +103,7 @@ def test_111fortranreaderbase(log, monkeypatch):
 
     Currently only tests logging functionality.
     '''
-    class FailFile(object):
+    class FailFile():
         '''
         A "file-like" object which returns a line of Fortran source followed
         by raising a StopIteration exception.
@@ -374,7 +374,7 @@ def check_include_works(fortran_filename, fortran_code, include_info,
     file specified by 'fortran_filename'.
     :param include_info: a list of 2-tuples each with an include \
     filename as a string followed by include code as a string.
-    :type include_info: list of (str, str)
+    :type include_info: List[str]
     :param str expected: the expected output after parsing the code.
     :param str tmpdir: the temporary directory in which to create and \
     process the Fortran files.
@@ -938,8 +938,7 @@ end python module foo
         string_pyf, ignore_comments=False)
     assert reader.format.mode == 'pyf', repr(reader.format.mode)
     for item in reader:
-        # Remove 'u's to allow for py2/3 unicode differences
-        assert re.sub("u", "", str(item)) == re.sub("u", "", expected.pop(0))
+        assert str(item) == expected.pop(0)
 
 
 def test_fix90():