+++ src/fparser/one/tests/test_scripts.py
@@ -36,7 +36,6 @@
 
 '''
 
-from __future__ import print_function
 import os
 import sys
 from fparser.scripts import parse