+++ src/fparser/two/tests/fortran2003/test_io_control_spec_r913.py
@@ -38,7 +38,6 @@ the list variant - r913.
 
 '''
 
-from __future__ import print_function
 import pytest
 from fparser.two import Fortran2003
 from fparser.two.parser import ParserFactory